+++ Sources/PatchDay/Settings/Alerts/DeliveryMethodMutationAlert.swift
@@ -11,7 +11,6 @@ class DeliveryMethodMutationAlert: PDAlert {
 
     private var sdk: PatchDataSDK?
     private let tabs: TabReflective?
-
     private let originalDeliveryMethod: DeliveryMethod
     private let originalQuantity: Int
     private let newDeliveryMethod: DeliveryMethod
@@ -24,22 +23,23 @@ class DeliveryMethodMutationAlert: PDAlert {
     }()
 
     lazy var continueHandler = {
-        if let sdk = self.sdk {
-            sdk.settings.setDeliveryMethod(to: self.newDeliveryMethod)
-            let defaultQuantity = DefaultQuantities.Hormone[self.newDeliveryMethod]
-            sdk.settings.setQuantity(to: defaultQuantity)
-        }
+        guard let sdk = self.sdk else { return }
+        sdk.settings.setDeliveryMethod(to: self.newDeliveryMethod)
         self.tabs?.reflectHormones()
     }
 
     lazy var declineAction: UIAlertAction = {
         UIAlertAction(title: ActionStrings.Decline, style: .cancel) {
-            _ in self.handlers.handleDecline(
-                originalMethod: self.originalDeliveryMethod,
-                originalQuantity: self.originalQuantity
-            )
+            _ in self.declineHandler()
         }
     }()
+
+    lazy var declineHandler = {
+        self.handlers.handleDecline(
+            originalMethod: self.originalDeliveryMethod,
+            originalQuantity: self.originalQuantity
+        )
+    }
 
     init(
         sdk: PatchDataSDK?,