+++ Tests/PatchDayTests/Integration/IntegrationTests.swift
@@ -33,6 +33,10 @@ class IntegrationTests: XCTestCase {
         let badge = PDBadge(sdk: sdk)
         sdk.settings.setDeliveryMethod(to: .Patches)  // Should trigger reset to 3 patches
         let ids = self.sdk.hormones.all.map({ $0.id })
+        if ids.count < 3 {
+            XCTFail("Hormone count did does not match delivery method")
+            return
+        }
         sdk.hormones.setDate(by: ids[0], with: DateFactory.createDate(daysFromNow: -20)!)
         sdk.hormones.setDate(by: ids[1], with: DateFactory.createDate(daysFromNow: -20)!)
         sdk.hormones.setDate(by: ids[2], with: DateFactory.createDate(daysFromNow: -20)!)