+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 3.3.2
+
+Fixed
+
+- Bug when setting the Delivery Method for the first time in an empty schedule would not adjust the Quantity
+    correctly.
+- Bug where you would receive a warning about the loss of data when changing the Delivery Method to the same 
+    setting that is already set.
+
 ## 3.3.1
 
 Fixed